pyproject.toml
[project]
name = "extractllm-aianalizy"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
